[refactor][rngrhn4114]: 본래 폴더명으로 수정
- rename folder temp-folder to solid-button
- rename folder temp-folder2 to text-button
- rename folder icon-folder to icon
